primitives: move Witness, Wtxid & WitnessScript into segwit module
